# Conflicts:
#	app/objectbox-models/default.json
#	app/objectbox-models/default.json.bak

+++ blecomm/src/main/java/com/microtechmd/blecomm/parser/CgmHistoryEntity.java
@@ -3,7 +3,7 @@ package com.microtechmd.blecomm.parser;
 public interface CgmHistoryEntity {
     void _setDatetime(long datetime);
     void _setEventIndex(int eventIndex);
-    void _setSensorIndex(long sensorIndex);
+    void _setSensorIndex(int sensorIndex);
     void _setEventType(int eventType);
     void _setEventValue(float eventValue);
     void _setRawValue(float[] rawValue);
